services/nillion/requirements.txt
python-dotenv==1.0.0
py-nillion-client
nada-dsl
nillion-python-helpers
python-dotenv==1.0.0
pytest-asyncio>=0.23.6
Flask[async]==3.0.3
flask-cors==4.0.0
cosmpy>=0.9.2